+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to `Muzzle` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [0.1.2] - 2018-06-21
+## [0.1.2] - 2018-06-22
 
 ### Added
 - route pattern matching on path assertions
+- improved `Transactions` usability
+- added shortcut methods to first and last request on the `Muzzle` instance
 
 ### Deprecated
 - Nothing